fix(forms): Move specific serialization logic to parent class and improve default itemtype
